krackattack/requirements.txt
pycryptodome==3.9.9
scapy==2.4.4
